test(checker): split assignability query-boundary tests
